[REF] Change type of the offset to off_t
When accessing storage we need to be able to address even large storages, so
the offset must be of an off_t type.

Fixes:
SEOS-1472

+++ components/log_clients/FileReaderWriter/FileReaderWriter.c
@@ -60,7 +60,7 @@ void post_init()
 int run(void)
 {
     OS_FileSystem_Handle_t hFs;
-    size_t length;
+    off_t length;
 
     /*************/
     /* Create fs */
